desktop/requirements.txt
# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
easyocr>=1.6.0
openai>=1.0.0
Pillow>=9.5.0
numpy>=1.21.0
matplotlib>=3.5.0
tqdm>=4.62.0

# GUI dependencies
tkinterdnd2>=0.4.3

# Development dependencies
pyinstaller>=5.0.0
black>=22.0.0